Compile method (#5)
* Naming change

* working compile method checkpoint

* Loading and saving

* Test commit

* updates to compile ux

* adding logging, todo for compile

* Finishing compile method - done main + some cleanup

+++ dagent/src/dagent/DecisionNode.py
@@ -1,41 +1,47 @@
 import json
+import os
 import inspect
 from .DagNode import DagNode
+
 from .base_functions import call_llm_tool, create_tool_desc
 from icecream import ic
+import logging
 
 
-class AgentNode(DagNode):
+class DecisionNode(DagNode):
     def __init__(
         self, 
-        func: callable, 
+        func: callable = call_llm_tool, 
         next_nodes: dict[str, DagNode] = None,
         user_params: dict | None = None  
     ):
         super().__init__(func, next_nodes)
         self.user_params = user_params or {}
+        self.logger = logging.getLogger(__name__)
+    
 
-    def compile(self, model='gpt-4-0125-preview') -> None:
-        #TODO: Test
+    def compile(self, model='gpt-4-0125-preview', force_load=False) -> None:
         """
         TODO
             - Add schema validation
             - Retry upon failure for generating tool description
             - Add error handling
-            - Saving the generated tool description, loading if exists 
+            - Code changing, updating tool description -> automatic?
+            - data models for passing info between nodes
         """
         for _, next_node in self.next_nodes.items():
             func_name = next_node.func.__name__ + '.json'
-            try:
-                with open(func_name, 'r') as f:
-                    data = json.load(f)
-            except FileNotFoundError:
+            self.logger.info(f"Compiling tool description for function: {next_node.func.__name__}")
+            if force_load or not os.path.exists(func_name):
                 tool_desc = create_tool_desc(model=model, function_desc=inspect.getsource(next_node.func))
-                data = json.loads(tool_desc)
+                tool_desc_json = json.loads(tool_desc)
                 with open(func_name, 'w') as f:
-                    json.dump(data, f)
+                    json.dump(tool_desc_json, f)
+            else:
+                with open(func_name, 'r') as f:
+                    tool_desc_json = json.load(f)
 
-            next_node.tool_description = json.loads(data)
+            next_node.tool_description = tool_desc_json 
             # TODO: rm
             ic(next_node.tool_description)
             next_node.compile()
@@ -44,6 +50,9 @@ class AgentNode(DagNode):
     def run(self, **kwargs) -> any:
         if not self.next_nodes:
             raise ValueError("Next nodes not specified for LLM call")
+        
+        if 'model' not in kwargs:
+            kwargs['model'] = 'gpt-4-0125-preview'
 
         try:
             # TODO: Messages param is unclear here to be passed
